refactor: Move functions
